pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pakunoda-mcp"
version = "0.1.0"
description = "MCP server exposing Pakunoda project state to AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0",
    "pyyaml>=6.0",
]

[project.scripts]
pakunoda-mcp = "pakunoda_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/pakunoda_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
